Rearranged class structure after watching video explaination of project
This commit has some pretty big changes. I rewatched the video and
realized that column C in the spreadsheet actually refers to how long a
gizmolio takes to be created. I also saw that the columns talking about
the start and end times and dates are not inputs, those are values are
program makes based off of when our algorithm runs it. Sorry I did not
catch this before we had to turn in checkpoint B. In addition I have
also created the SpreadsheetProcessing class so we can start working on
the selection algorithm before the database is implemented. As of now it
loads in the spreadsheet (must be converted to .csv) and stores all data
correctly.