Created GenerateOrder class to handle order generation and manufacturing
This class is the framework for generating orders and sending them to
the machine. Once the database is running all we have to do is some
minor changes to have the two work nicely together. There is only one
method that needs to be called, checkHourly(), and this takes care of
all of the logic and other methods from the class. The Javadocs goes
into further details.